feat(types): expand integer helpers for all 8 variants; Value::Null for null; Bool for compare
- runtime_integer_kind/value now cover I8/I16/I32/I64/U8/U16/U32/U64
- integer_value_from_kind produces distinct Value variants per kind (no more I8/I16/I32 → I64 collapse)
- check_integer_range returns i128 instead of i64 to avoid lossy cast
- runtime_null() returns Value::Null; expect_pointer coerces Null → null RawPointer for pointer ops
- runtime_compare returns Value::Bool; runtime_is_false handles Bool/Null/Unit/all 8 int zeros/floats
- is_null builtin handles Value::Null directly
- Update test expectations to reflect true/false output from Bool comparisons

+++ Rust/src/runtime/typing.rs
@@ -316,7 +316,7 @@ pub fn integer_range(kind: TypeKind) -> Option<(i128, i128)> {
     })
 }
 
-pub fn check_integer_range(kind: TypeKind, value: i128) -> Result<i64> {
+pub fn check_integer_range(kind: TypeKind, value: i128) -> Result<i128> {
     let (lo, hi) = integer_range(kind).ok_or_else(|| {
         TinyOneError::runtime(format!(
             "{:?} is not an integer type with a defined range",
@@ -330,7 +330,7 @@ pub fn check_integer_range(kind: TypeKind, value: i128) -> Result<i64> {
             kind.name()
         )));
     }
-    Ok(value as i64)
+    Ok(value)
 }
 
 #[cfg(test)]
@@ -431,8 +431,8 @@ mod tests {
     #[test]
     fn range_check_reports_memory_overflow() {
         assert!(check_integer_range(TypeKind::U8, 256).is_err());
-        assert_eq!(check_integer_range(TypeKind::U8, 255).unwrap(), 255);
+        assert_eq!(check_integer_range(TypeKind::U8, 255).unwrap(), 255i128);
         assert!(check_integer_range(TypeKind::I8, 128).is_err());
-        assert_eq!(check_integer_range(TypeKind::I8, -128).unwrap(), -128);
+        assert_eq!(check_integer_range(TypeKind::I8, -128).unwrap(), -128i128);
     }
 }
